Article store branch

+++ tango_with_django_project/rango/views.py
@@ -336,14 +336,15 @@ def visitor_cookie_handler(request):
     request.session['visits'] = visits
 
 
-
+@login_required
 def add_article(request):
-    if request.user.is_authenticated:
+    if request.user.is_authenticated:        
         if request.method == "POST":
             article_form = ArticleForm(request.POST)
             
             if article_form.is_valid():
-                data = article_form.save()
+                data = article_form.save(commit=False)
+                data.author = request.user
                 data.save()
                 
                 data_picture = article_form.save(commit=False)
@@ -364,46 +365,57 @@ def add_article(request):
     return redirect(reverse('rango:index'))
 
 
+
+@login_required
 def edit_article(request, pk):
     if request.user.is_authenticated:
         
         article = Article.objects.get(id=pk)
         form = ArticleForm(instance=article)
-        
-        if request.method == 'POST':
-            form = ArticleForm(request.POST, instance=article)
+        if request.user == article.author:
+            if request.method == 'POST':
+                form = ArticleForm(request.POST, instance=article)
+                if form.is_valid():
+                    form.save()
+                    
+                    data_picture = form.save(commit=False)
+                    data_picture.picture = form
+                    
+                    if 'picture' in request.FILES:
+                        data_picture.picture = request.FILES['picture']
+                    data_picture.save()
+                    
+                    return redirect('/any/article/')
 
-            if form.is_valid():
-                form.save()
-                
-                data_picture = form.save(commit=False)
-                data_picture.picture = form
-                
-                if 'picture' in request.FILES:
-                    data_picture.picture = request.FILES['picture']
-                data_picture.save()
-                
-                return redirect('/any/article/')
-        
-        context = {'form':form}
+            context = {'form':form}
+            
+        else:
+            return redirect('/any/article/')
     return render(request, 'rango/edit_article.html', context)
 
+@login_required
 def delete_article(request, pk):
     if request.user.is_authenticated:
         article = Article.objects.get(id=pk)
+        if request.user == article.author:
+            if request.method == "POST":
+                article.delete()
+                return redirect('/any/article/')
+            
+            context = {'item':article}
+            
         
-        if request.method == "POST":
-            article.delete()
+        else:
             return redirect('/any/article/')
-        context = {'item':article}
-        
+ 
     return render(request, 'rango/delete_article.html', context)
 
-
+@login_required
 def article_show(request):
-    article_list = Article.objects.all()
-    context_dict = {}
-    context_dict['articles'] = article_list
+    if request.user.is_authenticated:
+        article_list = Article.objects.all()
+        context_dict = {}
+        context_dict['articles'] = article_list
     
     return render(request, 'rango/article.html',context=context_dict)
